+++ src/router/router-guards/itemCommon.js
@@ -8,25 +8,20 @@ const { state } = store;
  * and their sub-pages.
  */
 export default async function itemCommon(to, from, next) {
+  if (!state.user) {
+    return next({
+      name: 'Login',
+      query: { redirectFrom: to.fullPath },
+    });
+  }
+
   const { activeItem } = state;
   const { slug } = to.params;
+  const slugRef = await getSlugRef(slug);
 
-  try {
-    const slugRef = await getSlugRef(slug);
-
-    if (!activeItem || !slugRef || activeItem.id !== slugRef.id) {
-      await store.dispatch('set_active_item', slugRef);
-    }
-
-    next();
-  } catch ({ message }) {
-    if (!state.user) {
-      next({
-        name: 'Login',
-        query: { redirectFrom: to.fullPath },
-      });
-    } else {
-      next({ name: 'Not found' });
-    }
+  if (!activeItem || !slugRef || activeItem.id !== slugRef.id) {
+    await store.dispatch('set_active_item', slugRef);
   }
+
+  return next();
 }