+++ src/views/Item/ItemMeasurements.vue
@@ -17,8 +17,13 @@
       </template>
     </template>
 
-    <template #default>
-      <kpi-details v-if="showKpiDetails && kpi" :kpi="kpi" />
+    <template v-if="showKpiDetails" #default>
+      <kpi-details v-if="kpi" :kpi="kpi" />
+      <not-found-page
+        v-else
+        :heading="$t('notFound.measurementHeading')"
+        :body="$t('notFound.measurementBody')"
+      />
     </template>
   </page-layout>
 
@@ -53,6 +58,7 @@ export default {
     KpiDetails,
     EmptyPage: () => import('@/components/pages/EmptyPage.vue'),
     PktButton: () => import('@oslokommune/punkt-vue2').then(({ PktButton }) => PktButton),
+    NotFoundPage: () => import('@/components/pages/NotFoundPage.vue'),
   },
 
   data: () => ({