+++ src/views/NotFound.vue
@@ -1,25 +1,19 @@
 <template>
-  <empty-page :heading="$t('404.title')" :body="$t('404.body')" skin="warning">
-    <router-link :to="{ name: 'Home' }" class="pkt-link">
-      <pkt-icon class="pkt-link__icon" name="chevron-left" />
-      {{ $t('404.linkText') }}
-    </router-link>
-  </empty-page>
+  <not-found-page back-to="Home" />
 </template>
 
 <script>
 import i18n from '@/locale/i18n';
-import EmptyPage from '@/components/pages/EmptyPage.vue';
 
 export default {
   name: 'NotFound',
   components: {
-    EmptyPage,
+    NotFoundPage: () => import('@/components/pages/NotFoundPage.vue'),
   },
 
   metaInfo() {
     return {
-      title: `${i18n.t('404.title')} | ${i18n.t('general.project')}`,
+      title: `${i18n.t('notFound.title')} | ${i18n.t('general.project')}`,
     };
   },
 };